+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.3.0] - 2021-06-17
+
+- Add permission calculator:
+
+```bash
+ac permissions convert --permission 640 [--symbolic|--octal]`
+```
+
 ## [0.2.4] - 2021-05-02
 
 - Refactor the git status to use a struct.